+++ protocol/synthetix/contracts/interfaces/IAssociateDebtModule.sol
@@ -11,24 +11,35 @@ import "@synthetixio/core-modules/contracts/interfaces/INftModule.sol";
 interface IAssociateDebtModule {
     /**
      * @notice Emitted when `associateDebt` is called.
+     * @param marketId The id of the market to which debt was associated.
+     * @param poolId The id of the pool associated to the target market.
+     * @param collateralType The address of the collateral type that acts as collateral in the corresponding pool.
+     * @param accountId The id of the account whose debt is being associated.
+     * @param amount The amount of debt being associated with the specified account.
+     * @param updateDebt The total updated debt of the account.
      */
     event DebtAssociated(
         uint128 indexed marketId,
         uint128 indexed poolId,
         address indexed collateralType,
         uint128 accountId,
-        uint amount,
-        int updatedDebt
+        uint256 amount,
+        int256 updatedDebt
     );
 
     /**
      * @notice Allows a market to associate debt with a specific position.
+     * @param marketId The id of the market to which debt was associated.
+     * @param poolId The id of the pool associated to the target market.
+     * @param collateralType The address of the collateral type that acts as collateral in the corresponding pool.
+     * @param accountId The id of the account whose debt is being associated.
+     * @param amount The amount of debt being associated with the specified account.
      */
     function associateDebt(
         uint128 marketId,
         uint128 poolId,
         address collateralType,
         uint128 accountId,
-        uint amount
-    ) external returns (int);
+        uint256 amount
+    ) external returns (int256);
 }
