+++ protocol/synthetix/contracts/interfaces/ICollateralConfigurationModule.sol
@@ -10,11 +10,14 @@ import "../storage/CollateralConfiguration.sol";
 interface ICollateralConfigurationModule {
     /**
      * @notice Emitted when a collateral type’s configuration is created or updated.
+     * @param collateralType The address of the collateral type that was just configured.
+     * @param config The object with the newly configured details.
      */
     event CollateralConfigured(address indexed collateralType, CollateralConfiguration.Data config);
 
     /**
      * @notice Creates or updates the configuration for the given `collateralType`.
+     * @param config The CollateralConfiguration object describing the new configuration.
      *
      * Requirements:
      *
@@ -28,20 +31,28 @@ interface ICollateralConfigurationModule {
     /**
      * @notice Returns a list of detailed information pertaining to all collateral types registered in the system.
      * @dev Optionally returns only those that are currently enabled.
+     * @param hideDisabled Wether to hide disabled collaterals or just return the full list of collaterals in the system.
+     * @returns The list of collateral configuration objects set in the system.
      */
-    function getCollateralConfigurations(
-        bool hideDisabled
-    ) external view returns (CollateralConfiguration.Data[] memory collaterals);
+    function getCollateralConfigurations(bool hideDisabled)
+        external
+        view
+        returns (CollateralConfiguration.Data[] memory collaterals);
 
     /**
      * @notice Returns detailed information pertaining the specified collateral type.
+     * @param collateralType The address for the collateral whose configuration is being queried.
+     * @returns The configuration object describing the given collateral.
      */
-    function getCollateralConfiguration(
-        address collateralType
-    ) external view returns (CollateralConfiguration.Data memory collateral);
+    function getCollateralConfiguration(address collateralType)
+        external
+        view
+        returns (CollateralConfiguration.Data memory collateral);
 
     /**
      * @notice Returns the current value of a specified collateral type.
+     * @param collateralType The address for the collateral whose price is being queried.
+     * @returns The price of the given collateral.
      */
-    function getCollateralPrice(address collateralType) external view returns (uint);
+    function getCollateralPrice(address collateralType) external view returns (uint256);
 }
