+++ protocol/synthetix/contracts/interfaces/ICollateralModule.sol
@@ -10,34 +10,52 @@ import "../storage/CollateralConfiguration.sol";
 interface ICollateralModule {
     /**
      * @notice Emitted when `amount` of collateral of type `collateralType` is deposited to account `accountId` by `sender`.
+     * @param accountId The if of the account that deposited collateral.
+     * @param collateralType The address of the collateral that was deposited.
+     * @param tokenAmount The amount of collateral that was deposited, denominated in the token's native decimal representation.
+     * @param sender The address of the account that triggered the deposit.
      */
     event Deposited(
         uint128 indexed accountId,
         address indexed collateralType,
-        uint tokenAmount,
+        uint256 tokenAmount,
         address indexed sender
     );
 
     /**
      * @notice Emitted when `amount` of collateral of type `collateralType` is withdrawn from account `accountId` by `sender`.
+     * @param accountId The if of the account that withdrew collateral.
+     * @param collateralType The address of the collateral that was withdrawn.
+     * @param tokenAmount The amount of collateral that was withdrawn, denominated in the token's native decimal representation.
+     * @param sender The address of the account that triggered the withdrawal.
      */
     event Withdrawn(
         uint128 indexed accountId,
         address indexed collateralType,
-        uint tokenAmount,
+        uint256 tokenAmount,
         address indexed sender
     );
 
     /**
      * @notice Deposits `amount` of collateral of type `collateralType` into account `accountId`.
      * @dev Anyone can deposit into anyone's active account without restriction.
+     * @param accountId The id of the account that is making the deposit.
+     * @param collateralType The address of the token to be deposited.
+     * @param tokenAmount The amount being deposited, denominated in the token's native decimal representation.
      *
      * Emits a {CollateralDeposited} event.
      */
-    function deposit(uint128 accountId, address collateralType, uint tokenAmount) external;
+    function deposit(
+        uint128 accountId,
+        address collateralType,
+        uint256 tokenAmount
+    ) external;
 
     /**
      * @notice Withdraws `amount` of collateral of type `collateralType` from account `accountId`.
+     * @param accountId The id of the account that is making the withdrawal.
+     * @param collateralType The address of the token to be withdrawn.
+     * @param tokenAmount The amount being withdrawn, denominated in the token's native decimal representation.
      *
      * Requirements:
      *
@@ -46,51 +64,68 @@ interface ICollateralModule {
      * Emits a {CollateralWithdrawn} event.
      *
      */
-    function withdraw(uint128 accountId, address collateralType, uint tokenAmount) external;
+    function withdraw(
+        uint128 accountId,
+        address collateralType,
+        uint256 tokenAmount
+    ) external;
 
     /**
      * @notice Returns the total values pertaining to account `accountId` for `collateralType`.
+     * @param accountId The id of the account whose collateral is being queried.
+     * @param collateralType The address of the collateral type whose amount is being queried.
+     * @returns totalDeposited The total collateral deposited in the account.
+     * @returns totalAssigned The amount of collateral in the account that is delegated to pools.
+     * @returns totalLocked THe amount of collateral in the account that cannot currently be un-delegated from a pool.
      */
-    function getAccountCollateral(
-        uint128 accountId,
-        address collateralType
-    )
+    function getAccountCollateral(uint128 accountId, address collateralType)
         external
         view
         returns (
-            uint totalDeposited,
-            uint totalAssigned,
-            uint totalLocked
+            uint256 totalDeposited,
+            uint256 totalAssigned,
+            uint256 totalLocked
             //uint totalEscrowed
         );
 
     /**
      * @notice Returns the amount of collateral of type `collateralType` deposited with account `accountId` that can be withdrawn or delegated to pools.
+     * @param accountId The id of the account whose collateral is being queried.
+     * @param collateralType The address of the collateral type whose amount is being queried.
+     * @returns The amount of collateral that is available for withdrawal or delegation.
      */
-    function getAccountAvailableCollateral(
-        uint128 accountId,
-        address collateralType
-    ) external view returns (uint);
+    function getAccountAvailableCollateral(uint128 accountId, address collateralType)
+        external
+        view
+        returns (uint256);
 
     /**
      * @notice Clean expired locks from locked collateral arrays for an account/collateral type. It includes offset and items to prevent gas exhaustion. If both, offset and items, are 0 it will traverse the whole array (unlimited).
+     * @param accountId The id of the account whose locks are being cleared.
+     * @param collateralType The address of the collateral type to clean locks for.
+     * @param offset The index of the first lock to clear.
+     * @param items The number of locks to clear.
      */
     function cleanExpiredLocks(
         uint128 accountId,
         address collateralType,
-        uint offset,
-        uint items
+        uint256 offset,
+        uint256 items
     ) external;
 
     /**
      * @notice Create a new lock on the given account. you must have `admin` permission on the specified account to create a lock.
      * @dev A collateral lock does not affect withdrawals, but instead affects collateral delegation.
      * @dev There is currently no benefit to calling this function. it is simply for allowing pre-created accounts to have locks on them if your protocol requires it.
+     * @param accountId The id of the account for which a lock is to be created.
+     * @param collateralType The address of the collateral type for which the lock will be created.
+     * @param amount The amount of collateral tokens to wrap in the lock being created.
+     * @param expireTimestamp The date in which the lock will become clearable.
      */
     function createLock(
         uint128 accountId,
         address collateralType,
-        uint amount,
+        uint256 amount,
         uint64 expireTimestamp
     ) external;
 }
